// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  STUDENT: 'STUDENT',
  TEACHER: 'TEACHER',
  STAFF: 'STAFF',
  ALUMNI: 'ALUMNI',
  OTHER: 'OTHER',
  ADMIN: 'ADMIN'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const WorkStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED',
  ARCHIVED: 'ARCHIVED'
} as const

export type WorkStatus = (typeof WorkStatus)[keyof typeof WorkStatus]


export const Visibility = {
  PUBLIC: 'PUBLIC',
  SCHOOL: 'SCHOOL',
  PRIVATE: 'PRIVATE'
} as const

export type Visibility = (typeof Visibility)[keyof typeof Visibility]


export const EventType = {
  EXHIBITION: 'EXHIBITION',
  RECRUITMENT: 'RECRUITMENT',
  CONTEST: 'CONTEST',
  OTHER: 'OTHER'
} as const

export type EventType = (typeof EventType)[keyof typeof EventType]


export const NotificationType = {
  SYSTEM: 'SYSTEM',
  LIKE: 'LIKE',
  COMMENT: 'COMMENT',
  EVENT: 'EVENT'
} as const

export type NotificationType = (typeof NotificationType)[keyof typeof NotificationType]